+++ src/distribution.ts
@@ -2,15 +2,15 @@ import type { Socket } from "socket.io";
 import { SomeRemoteSocket } from "./remote-client";
 
 export interface Distribution {
-  join: (rooms: string | string[]) => void | Promise<void>;
-  leave: (rooms: string | string[]) => void | Promise<void>;
+  join: (rooms: string | string[]) => Promise<void>;
+  leave: (rooms: string | string[]) => Promise<void>;
 }
 
 export const makeDistribution = (
   subject: Socket | SomeRemoteSocket,
 ): Distribution => ({
-  join: subject.join.bind(subject),
-  leave: (rooms) =>
+  join: async (rooms) => subject.join(rooms),
+  leave: async (rooms) =>
     typeof rooms === "string"
       ? subject.leave(rooms)
       : Promise.all(rooms.map((room) => subject.leave(room))).then(() => {}),