+++ src/index.spec.ts
@@ -1,6 +1,8 @@
+import type { IncomingMessage } from "node:http";
 import { expectNotType, expectType } from "tsd";
 import { z } from "zod";
 import {
+  AbstractLogger,
   ClientContext,
   EmissionMap,
   LoggerOverrides,
@@ -15,6 +17,7 @@ describe("Entrypoint", () => {
   });
 
   test("should expose certain types and interfaces", () => {
+    expectType<AbstractLogger>(console);
     expectType<EmissionMap>({});
     expectType<EmissionMap>({ event: { schema: z.tuple([]) } });
     expectType<EmissionMap>({
@@ -31,6 +34,7 @@ describe("Entrypoint", () => {
     >({
       emission: { event: { schema: z.tuple([]) } },
       hooks: {},
+      examples: {},
       metadata: z.object({ count: z.number() }),
     });
     expectType<
@@ -41,6 +45,7 @@ describe("Entrypoint", () => {
     >({
       client: {
         isConnected: () => true,
+        getRequest: <T>() => ({}) as IncomingMessage as T,
         id: "",
         handshake: {
           headers: {},
@@ -55,8 +60,8 @@ describe("Entrypoint", () => {
         },
         getData: () => ({ count: 1 }),
         setData: () => {},
-        join: () => {},
-        leave: () => {},
+        join: async () => {},
+        leave: async () => {},
         getRooms: () => [""],
         emit: async () => true,
         broadcast: async () => true,
@@ -81,8 +86,8 @@ describe("Entrypoint", () => {
             rooms: [""],
             getData: () => ({ count: 1 }),
             emit: async () => true,
-            join: () => {},
-            leave: () => {},
+            join: async () => {},
+            leave: async () => {},
           },
         ],
         broadcast: async () => true,